Rename "Regression" → "Comparison"
