+++ tools/tck-report-comparison/README.adoc
@@ -11,7 +11,7 @@ https://github.com/junit-team/junit5/issues/990[available before JUnit 5.1.1 ver
 and is https://issues.apache.org/jira/browse/SUREFIRE-1546[fixed in Surefire 3.0.0-M4] (not released yet).
 If version change is not possible - use link:../tck-reporting[Cucumber report].
 
-Comparsion report in HTML format or CSV file will be generated in
+Comparison report in HTML format or CSV file will be generated in
 `build/reports/tests/` directory. To change directory use
 `--output /path/to/dir`.
 
@@ -48,15 +48,15 @@ Verify that all failing scenarios in `report1` are in `report2`.
 Fail the build if there are new passing or failing scenarios in
 `report1`.
 
-== Create CSV from XML report
+== Create CSV from XML or JSON report
 
-Storing JUnit XML report in the repository might not be practical, as it
+Storing JUnit XML or Cucumber reports in the repository might not be practical, as it
 contains a lot of surplus information. CSV with format
 `feature,scenario,status` might be more manageable.
 
-Arguments: `/path/to/report1.xml --saveCsv`
+Arguments: `/path/to/report1 --saveCsv`
 
-Will convert `report1.xml` to `base.csv` in output directory. Resulting CSV will have following columns:
+Will convert `report1` to `base.csv` in output directory. Resulting CSV will have following columns:
 `feature,scenario,status`.
 
 == Usage
@@ -68,12 +68,17 @@ Run from command line:
 [source,sh]
 ----
 mvn clean install -DskipTests
-mvn -pl tools/tck-regression exec:java -Dexec.mainClass="org.opencypher.tools.tck.regression.TckResultsComparator"
+mvn -pl tools/tck-report-comparison exec:java -Dexec.mainClass="org.opencypher.tools.tck.reporting.comparison.TckResultsComparator"
     -Dexec.args="<arguments>"
-# Example
-mvn -pl tools/tck-regression exec:java -Dexec.mainClass="org.opencypher.tools.tck.regression.TckResultsComparator"
+# Example for JUnit xml
+mvn -pl tools/tck-report-comparison exec:java -Dexec.mainClass="org.opencypher.tools.tck.reporting.comparison.TckResultsComparator"
     -Dexec.args="/path/to/report1.xml --compareTo /path/to/report2.xml"
+# Example for Cucumber json
+mvn -pl tools/tck-report-comparison exec:java -Dexec.mainClass="org.opencypher.tools.tck.reporting.comparison.TckResultsComparator"
+    -Dexec.args="/path/to/cucumber.json --compareTo /path/to/cucumber.json"
 ----
+
+Report type is detected by extension. Supported extensions: `.xml`, `.json`, `.csv`.
 
 === Integrate in build
 
@@ -95,7 +100,7 @@ https://www.mojohaus.org/exec-maven-plugin/usage.html[exec-maven-plugin]
             <goal>java</goal>
           </goals>
           <configuration>
-            <mainClass>org.opencypher.tools.tck.regression.TckResultsComparator</mainClass>
+            <mainClass>org.opencypher.tools.tck.reporting.comparison.TckResultsComparator</mainClass>
             <arguments>
               <argument>/path/to/report1.xml</argument>
               <argument>--verifyFailures</argument>
