+++ tools/tck-report-comparison/pom.xml
@@ -10,8 +10,8 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>tck-regression</artifactId>
-    <name>openCypher TCK Regression</name>
+    <artifactId>tck-report-comparison</artifactId>
+    <name>openCypher TCK Report Comparison</name>
     <url>http://opencypher.org</url>
     <description>Compare TCK JUnit reports</description>
